Extract new interface to a header file
For now, all the code is put in a single header file. Later on it should
be split into logical units.

Closes #75
